chat/api/serializers.py
from rest_framework import serializers
from django.db import models
from chat.models import Chat, Contact
from chat.views import get_user_contact
from django.contrib.auth.models import User


class ContactSerializer(serializers.StringRelatedField):
    def to_internal_value(self, value):
        return value


class ChatSerializer(serializers.ModelSerializer):
    participants = ContactSerializer(many=True)

    class Meta:
        model = Chat
        fields = ('id', 'messages', 'participants')
        read_only = ('id')

    def create(self, validated_data):
        participants = validated_data.pop('participants')
        chat = Chat()
        chat.save()
        for username in participants:
            u = User.objects.get(username=username)
            if Contact.objects.filter(user=u).count() == 0:
                Contact.objects.create(
                    user=u
                )

            contact = get_user_contact(username)
            chat.participants.add(contact)
        chat.save()
        return chat


# do in python shell to see how to serialize data

# from chat.models import Chat
# from chat.api.serializers import ChatSerializer
# chat = Chat.objects.get(id=1)
# s = ChatSerializer(instance=chat)
# s
# s.data
