requirements.txt
aioredis==1.2.0
apt-xapian-index==0.51
apturl==0.5.2
asgiref==2.3.2
asn1crypto==0.24.0
astroid==2.0.4
async-timeout==3.0.1
attrs==18.2.0
autobahn==18.10.1
Automat==0.7.0
autopep8==1.5
beautifulsoup4==4.6.0
Brlapi==0.6.6
certifi==2018.10.15
cffi==1.14.0
channels==2.1.5
channels-redis==2.3.1
chardet==3.0.4
command-not-found==0.3
constantly==15.1.0
cryptography==2.8
cupshelpers==1.0
daphne==2.2.2
defer==1.0.6
defusedxml==0.5.0
distro-info===0.18ubuntu0.18.04.1
Django==2.2
django-allauth==0.38.0
django-cors-headers==2.4.0
django_heroku=0.3.1
django-rest-auth==0.9.3
django-rest-framework==0.1.0
djangorestframework==3.9.0
gunicorn==19.7.1
hiredis==0.2.0
html5lib==0.999999999
httplib2==0.9.2
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
isort==4.3.4
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.3.1
louis==3.5.0
lxml==4.2.1
macaroonbakery==1.1.3
mccabe==0.6.1
msgpack==0.5.6
netifaces==0.10.4
oauth==1.0.1
oauthlib==2.1.0
olefile==0.45.1
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.0.0
psycopg2-binary==2.7.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.5.0
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
PyHamcrest==1.9.0
pylint==2.1.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyOpenSSL==19.1.0
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.6.1
python-debian==0.1.32
python3-openid==3.1.0
pytz==2018.7
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.20.1
requests-oauthlib==1.0.0
requests-unixsocket==0.1.5
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.13.2
six==1.11.0
sqlparse==0.3.0
system-service==0.3
systemd-python==234
Twisted==18.9.0
txaio==18.8.1
typed-ast==1.1.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.2
usb-creator==0.3.3
virtualenv==15.1.0
wadllib==1.3.2
webencodings==0.5
whitenoise==3.3.1
wrapt==1.10.11
xkit==0.0.0
zope.interface==4.6.0
